chore: Moved the webhook handler methods to core
